Switch `ctr` to `ctb` in authorship.
`Contributor [ctb]`
A person, family or organization responsible for making contributions to the resource. This includes those whose work has been contributed to a larger work, such as an anthology, serial publication, or other compilation of individual works. If a more specific role is available, prefer that, e.g. editor, compiler, illustrator
UF Collaborator

`Contractor [ctr]`
A person or organization relevant to a resource, who enters into a contract with another person or organization to perform a specific task.

https://www.loc.gov/marc/relators/relaterm.html